+++ CHANGELOG.md
@@ -15,10 +15,8 @@
 - [6126](https://github.com/vegaprotocol/vega/issues/6126) - Don't generate market depth subscription messages if nothing has changed
 
 ### 🐛 Fixes
-- [6296](https://github.com/vegaprotocol/vega/issues/6296) - Add `assetId` as a field to the `GraphQL UpdateAsset` type
-- [6324](https://github.com/vegaprotocol/vega/issues/6324) - Ensure all `datanode` insert points use dedicated connection
-- [6273](https://github.com/vegaprotocol/vega/issues/6273) - Deposits return an empty list when no party is provided
 - [6307](https://github.com/vegaprotocol/vega/issues/6307) - Emit an event with status rejected if a protocol upgrade proposal has no validator behind it
+- [5305](https://github.com/vegaprotocol/vega/issues/5305) - Handle market updates changing price monitoring parameters correctly.
 
 ## 0.56.0